requirements.txt
asgiref==3.5.0
Django==3.2.13
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
